+++ core/digital-twin-lib/src/interfaces/DigitalTwinInterface.ts
@@ -1,9 +1,1 @@
-export interface DigitalTwinInterface {
-  address: string;
-  containers: any[];
-  created: number;
-  description: string;
-  name: string;
-  owner: string;
-  updated: number;
-}
+export { DigitalTwin as DigitalTwinInterface } from '@evan.network/api-blockchain-core';
