--- core/digital-twin-lib/src/interfaces/DigitalTwinTemplate.ts
@@ -1,20 +1,0 @@
-import { DBCPDescriptionInterface } from '../DAppContract';
-import { ContainerPlugin } from '@evan.network/api-blockchain-core';
-
-export interface DigitalTwinTemplateInterface {
-  description: DBCPDescriptionInterface;
-  plugins: { [pluginName: string]: ContainerPlugin };
-}
-
-export interface ValidationErrorInterface {
-  dataPath: string;
-  keyword: string;
-  message: string;
-  params: any;
-  schemaPath: string;
-}
-
-export interface TemplateErrorInterface {
-  name: string;
-  errors: ValidationErrorInterface[];
-}
