+++ core/digital-twin-lib/src/interfaces/TemplateValidationErrors.ts
@@ -1,0 +1,12 @@
+export interface ValidationErrorInterface {
+  dataPath: string;
+  keyword: string;
+  message: string;
+  params: any;
+  schemaPath: string;
+}
+
+export interface TemplateErrorInterface {
+  name: string;
+  errors: ValidationErrorInterface[];
+}
